// src/app/shared/enums/sweeAlert.enum.ts
export enum SWEET_ALERT_ICON {
  SUCCESS = 'success',
  ERROR = 'error',
  WARNING = 'warning',
  INFO = 'info',
  QUESTION = 'question',
}

export enum SWEET_ALERT_POSITION {
  TOP = 'top',
  TOP_START = 'top-start',
  TOP_END = 'top-end',
  TOP_LEFT = 'top-left',
  TOP_RIGHT = 'top-right',
  CENTER = 'center',
  CENTER_START = 'center-start',
  CENTER_END = 'center-end',
  CENTER_LEFT = 'center-left',
  CENTER_RIGHT = 'center-right',
  BOTTOM = 'bottom',
  BOTTOM_START = 'bottom-start',
  BOTTOM_END = 'bottom-end',
  BOTTOM_LEFT = 'bottom-left',
  BOTTOM_RIGHT = 'bottom-right',
}