// src/app/shared/models/user.model.ts
export class USER {
  nombreUsuario: string;
  apellidoUsuario: string;
  fechaNacimiento: string;
  correo: string;
  contrasenia: string;
  nombrePais: string;
  nombreIdiomaNativo: string;
  urlFoto: string;
  descripcion: string;
  nombreIdiomaAprendiz: string;
  nombreNivelIdiomaAprendiz: string;
  nombreIntereses: string[];
}